requirements.txt
pyodbc
fastapi
mysql-connector-python
python-dotenv
uvicorn
python-jose
bcrypt
passlib
python-multipart
google-generativeai